Removed MainActivity. Now our entry point is WordsActivity.
WordsActivity will now use a ViewPager and not create multiple instances of itself.

WordListType now needs a getTypeFromPosition() method.

WordsFragment will now figure out which set of words to show based on the configuration bundle obtained via getArgument() and not via the Intent extras of the parent activity.

Manifest updated to remove MainActivity.

Layout activity_words now is a ViewPager and not a LinearLayout.